pyproject.toml
[project]
name = "ai-design-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "haystack-ai>=2.21.0",
    "ipykernel>=7.1.0",
    "langchain-community>=0.4.1",
    "langchain[openai]>=1.2.0",
    "langsmith>=0.5.0",
    "pydantic>=2.12.5",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
]


[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"
